+++ src/math.ts
@@ -1,4 +1,3 @@
-import { Type } from "."
 import { HeaderTypes } from "./data/consts"
 
 /**
